chore(contract): cleanup migrations after 3.13.0 release
